Make the user and directory relative
ie not my user name and directory but a relative directory:
  ~/projects/failure-driven/rails-metal-cloud-metal/tmp/deploy
and the $USER of the machine
  CAP_DEPLOY_USER="${USER}" bundle exec cap production deploy

Would be still nicer to dictate where the code is deployed? not sure if
we can run capistrano without the SSH mode? at which stage switching to
something simpler might be better? but with SSH mode it will make the
next step easier to deploy to a separate *Nix box.